Navigation Upgrades
The navigation header works now. It does not go between pages, but
different 'pages' on one page. This is fairly easy to do. There is a
prototype of splitview functionality for moving between different types
of topics. The functionality has not been implemented but it will look
similar to what it does now.