Splitview Functionality
Splitview now has some basic functionality. We still need to set up some
specific pages for it to navigate to, but the navigation in general is
working. I added some styles to the app.xaml but they don't seem to be
doing much, will most likely remove/play around with in the future. I'm
leaving them there for now as it took me forever to figure out where you
add styles.