Automatic SplitView closing
If the mouse leaves the SplitView, the side bar will automatically
contract (a better version of this would be if the user clicks something
outside of the area it closes). The SplitView also now only opens to
width 150 to save space.

I also changed the colour of the SplitView to the Faculty of Computer
Science colour, which is given by the university. It doesn't look great
so we might want to add the colour elsewhere.